rename Registry -> ClassRegistry
