+++ src/example_app/comp/beans.py
@@ -1,4 +1,4 @@
-from jivago.inject.registry import Component, Singleton
+from jivago.inject.class_registry import Component, Singleton
 
 
 @Component
