+++ src/jivago/inject/annoted_class_binder.py
@@ -1,12 +1,12 @@
 from jivago.config import AbstractBinder
-from jivago.inject.registry import Registry, Annotation
+from jivago.inject.class_registry import ClassRegistry, Annotation
 from jivago.inject.service_locator import ServiceLocator
 from jivago.lang.annotations import Override
 
 
 class AnnotatedClassBinder(AbstractBinder):
 
-    def __init__(self, root_package: str, registry: Registry, annotation: Annotation):
+    def __init__(self, root_package: str, registry: ClassRegistry, annotation: Annotation):
         self.annotation = annotation
         self.rootPackage = root_package
         self.registry = registry
